+++ app/admin/events.rb
@@ -1,0 +1,3 @@
+ActiveAdmin.register Event do
+  permit_params :name, :url, :address, :country_code, :start_date, :end_date
+end