+++ app/models/event.rb
@@ -1,0 +1,7 @@
+class Event < ApplicationRecord
+  # :nocov:
+  def self.ransackable_attributes(_auth_object = nil)
+    %w[address country_code end_date name start_date url]
+  end
+  # :nocov:
+end